+++ com.unity.render-pipelines.high-definition/Editor/Material/Decal/ShaderGraph/DecalSubShader.cs
@@ -54,7 +54,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
             },
 
-            UseInPreview = true,
+            UseInPreview = false,
             OnGeneratePassImpl = (IMasterNode node, ref Pass pass) =>
             {
 
@@ -118,7 +118,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
             },
 
-            UseInPreview = true,
+            UseInPreview = false,
             OnGeneratePassImpl = (IMasterNode node, ref Pass pass) =>
             {
 
@@ -176,7 +176,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
             },
 
-            UseInPreview = true,
+            UseInPreview = false,
         };
 
 
@@ -229,7 +229,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
             },
 
-            UseInPreview = true,
+            UseInPreview = false,
             OnGeneratePassImpl = (IMasterNode node, ref Pass pass) =>
             {
 
@@ -300,7 +300,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             {
             },
 
-            UseInPreview = true,
+            UseInPreview = false,
             OnGeneratePassImpl = (IMasterNode node, ref Pass pass) =>
             {
 
@@ -335,6 +335,57 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             ZTestOverride = "ZTest LEqual",
             ZWriteOverride = "ZWrite Off",
             BlendOverride = "Blend 0 SrcAlpha One",
+
+            ExtraDefines = new List<string>()
+            {
+            },
+
+            Includes = new List<string>()
+            {
+                "#include \"Packages/com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/ShaderPass/ShaderPassDecal.hlsl\""
+            },
+
+            RequiredFields = new List<string>()
+            {
+                "AttributesMesh.normalOS",
+                "AttributesMesh.tangentOS",
+                "AttributesMesh.uv0",
+
+                "FragInputs.worldToTangent",
+                "FragInputs.positionRWS",
+                "FragInputs.texCoord0",
+            },
+
+            PixelShaderSlots = new List<int>()
+            {
+                DecalMasterNode.AlbedoSlotId,
+                DecalMasterNode.BaseColorOpacitySlotId,
+                DecalMasterNode.NormalSlotId,
+                DecalMasterNode.NormaOpacitySlotId,
+                DecalMasterNode.MetallicSlotId,
+                DecalMasterNode.AmbientOcclusionSlotId,
+                DecalMasterNode.SmoothnessSlotId,
+                DecalMasterNode.MAOSOpacitySlotId,
+                DecalMasterNode.EmissionSlotId
+            },
+
+            VertexShaderSlots = new List<int>()
+            {
+            },
+
+            UseInPreview = false,
+        };
+
+        // Pass to have a preview
+        Pass m_PassPreview = new Pass()
+        {
+            Name = "ForwardOnly",
+            LightMode = "ForwardOnly",
+            TemplateName = "DecalPass.template",
+            MaterialName = "Decal",
+            ShaderPassName = "SHADERPASS_FORWARD_PREVIEW",
+
+            ZTestOverride = "ZTest LEqual",
 
             ExtraDefines = new List<string>()
             {
@@ -480,6 +531,11 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 if (masterNode.affectsEmission.isOn)
                 {
                     GenerateShaderPass(masterNode, m_PassMeshEmissive, mode, subShader, sourceAssetDependencyPaths);
+                }
+
+                if (mode.IsPreview())
+                {
+                    GenerateShaderPass(masterNode, m_PassPreview, mode, subShader, sourceAssetDependencyPaths);
                 }
             }
             subShader.Deindent();